requirements.txt
kivy==2.3.1
kivymd==1.2.0
plyer==2.1.0
python-telegram-bot==22.2
requests==2.32.3
schedule==1.2.2
psutil==7.0.0
